feat(hooks): migrate useAccordionToggle to be named export
Also moves this inside AccordionToggle, as it makes sense to be
exporting from AccordionToggle rather than its own file.

Add an example import statement for useAccordionToggle in API
docs.

BREAKING CHANGE: useAccordionToggle is now being exported from
`AccordionToggle` rather than `useAccordionToggle`. It also needs
to be imported as a named import, rather than an unnamed import.
