--- types/components/useAccordionToggle.d.ts
@@ -1,5 +1,0 @@
-export default interface useAccordionToggle {
-  (eventKey: string, onClick: (event?: React.SyntheticEvent) => void): (
-    event?: React.SyntheticEvent,
-  ) => void;
-}
