+++ www/src/pages/components/accordion.mdx
@@ -53,6 +53,8 @@ You can now hook into the Accordion toggle functionality via `useAccordionToggle
 ### useAccordionToggle
 
 ```js
+import { useAccordionToggle } from 'react-bootstrap/AccordionToggle';
+
 const decoratedOnClick = useAccordionToggle(eventKey, onClick);
 ```
